requirements.txt
graphics.py
simpleaudio